Configure transitive Groovy 3 dependencies for Groovy tools
With older versions of Groovy, a custom-repackaged groovy-all dependency
was used as `localGroovy()` which had all the required tools. This is no
longer the case with Groovy 3 - we do not repackage an `-all` jar.
Thus, if Groovy 3 is found on the classpath (and current implementation
does not reveal if it came from an external dependency of from `localGroovy()`)
add external dependencies required by the tools.