Separate Groovydoc/compileGroovy classpaths with Groovy 3
Groovy 3 is not repackaged in groovy-all jar that would include
all the dependencies required for its tools. With Groovy 3,
when `localGroovy()` is used, the transitive dependencies
required for groovydoc are not on the classpath.
Those tools require an external repository in order to be resolved.

This change separates the classpaths for Groovy compiler and Groovydoc,
allowing `compileGroovy` to be executed when `localGroovy()` is used and
not requiring external repositories. Groovydoc task will fail unless
external repositories for transitive dependencies are defined.